Docstrings

There are two ways to use this package for LGF evaluation.

The high level interface performs LGF evaluations by automatically switching from quadrature to a series expansion whenever it is permissible within a given error tolerance. The user specifies the tolerance and the number of series terms used. This functionality is exported by the package.

The low level interface provides direct access to expansion coefficients, function generation, and error estimates for each type of series. This functionality is not exported.

High level interface

Stencil Types

ExpandLGF.Stencil — Type

An abstract type for all finite difference stencils

ExpandLGF.SplitStencil — Type

A dimension-split finite difference stencil

ExpandLGF.FullStencil — Type

A general symmetric finite difference stencil in 2D or 3D

ExpandLGF.MehrstellenStencil — Type

A Mehrstellen stencil with both left and right finite difference stencils

ExpandLGF.Mehrstellen4 — Function

Return the 4th order 3D Mehrstellen stencil

ExpandLGF.Mehrstellen6 — Function

Return the 6th order 3D Mehrstellen stencil

ExpandLGF.LeftMehrstellen4 — Function

Return the stencil for the 4th order left-hand side Mehrstellen operator

ExpandLGF.LeftMehrstellen6 — Function

Return the stencil for the 6th order left-hand side Mehrstellen operator

ExpandLGF.StandardDifference2D — Function

Return a 2D dimension-split stencil of given order and minimal width

ExpandLGF.StandardDifference3D — Function

Return a 3D dimension-split stencil of given order and minimal width

LGF Evaluation

ExpandLGF.EvaluateLGF — Type

A callable struct wrapping near-field evaluation, far-field evaluation, and a cutoff between the two. Call via (::EvaluateLGF)(n) for a 2D or 3D integer vector n.

ExpandLGF.EvaluateLGF — Method

EvaluateLGF(stencil::SplitStencil, near_terms, far_terms, rtol, atol)

Return a callable struct that evaluates the LGF by switching between near-field and far-field evaluation strategies.

Arguments

stencil : the split finite difference stencil

near_terms : number of inner and outer integral expansion terms.

far_terms : number of far-field expansion terms.

rtol, atol : relative and absolute error tolerance for the evaluation.

EvaluateLGF(stencil::Union{FullStencil, MehrstellenStencil}, far_terms, rtol, atol, maxevals)

Return a callable struct that evaluates the LGF for a FullStencil or MehrstellenStencil, switching between near-field or far-field evaluation strategies.

Arguments

stencil : the finite difference stencil

far_terms : number of far-field expansion terms.

rtol, atol : relative and absolute error tolerance for the evaluation.

maxevals : the max number of symbol evaluations used to compute near-field LGF values

ExpandLGF.NearFieldEvaluator — Type

A callable structure that evaluates near-field values through quadrature. Call with (::NearFieldEvaluator)(n) for an integer vector n.

ExpandLGF.NearFieldSplit — Type

A callable structure that evaluates near-field values for a dimension-split stencil

ExpandLGF.NearFieldFull — Type

A callable structure that evaluates near-field values via quadrature for a full or Mehrstellen stencil

ExpandLGF.NearField — Function

NearField(stencil::SplitStencil, nmax, terms, rtol, atol)

Return a callable struct which evaluates the LGF for a SplitStencil for $\|n\|_\infty \le {n_\mathrm{max}}$ with prescribed error tolerance. Uses a combination of direct quadrature and series expansions.

Arguments

stencil : the finite difference stencil

nmax : the largest component of n that will be evaluated

terms : number of terms used in integral expansions. Affects performance.

rtol, atol : relative and absolute error tolerance for LGF values

NearField(stencil::Stencil, rtol, atol, maxevals)

Return a callable struct which evaluates the LGF for any Stencil with prescribed error tolerance. Uses direct quadrature.

Arguments

stencil : the finite difference stencil

nmax : the largest component of n that will be evaluated

rtol, atol : relative and absolute error tolerance for LGF values

maxevals : the max number of symbol evaluations used to compute each LGF value

ExpandLGF.FarFieldExpansion — Type

A callable struct for far field evaluation. Call via (::FarFieldExpansion)(n) for a 2D or 3D integer vector n.

ExpandLGF.FarField — Function

far_field FarField(stencil::Stencil, nterms)

Evaluate the LGF for stencil via a far-field series expansion with ntemrs terms.

far_field, nmin = FarField(stencil::Stencil, nterms, rtol, atol)

Evaluate the LGF for $\|n\|_2 \ge n_{\mathrm{min}}$ via series expansion with prescribed error tolerance. Returns a callable struct and the threshold nmin beyond which the expansion achieves the given tolerance.

Arguments

stencil - the finite difference stencil

dimensions : either 2 or 3 to indicate a 2D or 3D LGF

nterms - number of terms in the expansion

rtol, atol - relative and absolute error tolerance on the evaluation (pass 0 to ignore either).

Low level interface

Expansion coefficients

ExpandLGF.inner_integral_expansion_coeffs — Function

inner_integral_expansion_coeffs(stencil::SplitStencil, n, m)

Return an $m$-term expansion of the inner integral for fixed $n$ and large $t$. The results is an array b of univariate polynomials in n so that

\[I_{\sigma, n}(t) = \frac{1}{\sqrt{4\pi t}} \sum_{k=0}^m b_k(n) t^{-k} + \mathcal{O}\left( t^{-\frac{2m+3}{2}} \right)\]

Arguments

stencil : the finite difference stencil

m : the number of terms in the expansion

PR : a univariate PolynomialRing

ExpandLGF.outer_integral_expansion_coeffs — Function

outer_integral_expansion_coeffs(stencil::SplitStencil{3, T}, n, m) where T

Return an $m$-term expansion of the 3D outer integral for large $T$. The results is an array $g$ of mulitvariate polynomials in n so that

\[\int_T^\infty I_{\sigma, n_1}(t) I_{\sigma, n_2}(t) I_{\sigma, n_3}(t) \,\mathrm{d}t \approx \frac{1}{\sqrt{16\pi^3T}} \sum_{k=0}^m g_k(\bm{n}) T^{-k} + \mathcal{O}\left( T^{-\frac{2m+3}{2}} \right)\]

Arguments

stencil : the finite difference stencil

n : a vector of 3 generators from a multivariate PolynomialRing

m : the number of terms in the expansion

outer_integral_expansion_coeffs(stencil::SplitStencil{2, T}, n, m) where T

Return an $m$-term expansion of the 2D outer integral for large $T$. The results is an array $g$ of mulitvariate polynomials in n so that

\[\int_T^\infty [I_{\sigma, n_1}(t) I_{\sigma, n_2}(t) - I_{\sigma, 0}(t)^2] \,\mathrm{d}t \approx \frac{1}{4\pi} \sum_{k=1}^m g_k(\bm{n}) T^{-k} + \mathcal{O}\left( T^{-(m+1)} \right)\]

Arguments

stencil : the finite difference stencil

n : a vector of 2 generators from a multivariate PolynomialRing

m : the number of terms in the expansion

ExpandLGF.outer_integral_2d_expansion_coeffs — Function

outer_integral_2d_expansion_coeffs(b, PR::MPolyRing)

Return a power series expansion of the 2D outer integral based on an expansion of the inner integral.

Arguments

b : inner expansion, the result of inner_integral_expansion_coeffs()

PR : a multivariate PolynomialRing in two variables

ExpandLGF.outer_integral_3d_expansion_coeffs — Function

outer_integral_3d_expansion_coeffs(b, n)

Return a power series expansion of the 3D outer integral based on an expansion of the inner integral.

Arguments

b : inner expansion, the result of inner_integral_expansion_coeffs()

PR : a multivariate PolynomialRing with 3 generators

ExpandLGF.graded_series_expansion — Function

bj = graded_series_expansion(stencil::Stencil{T}, j_max::Int, PR::MPolyRing{T}) where T

Return a power series expansion of the Fourier symbol for a given stencil:

\[\sigma(\bm{k}) = \sum_{j = 0}^{j_{\mathrm{max}}} b_j(\bm{k}) + \mathcal{O}\left( |\bm{k}|^{2j_{\mathrm{max}} + 2} \right)\]

In the output, b[j] contains a homogeneous multivariate polynomial of degree 2j.

Arguments

stencil : the stencil defining the symbol

j_max : the number of valid entries in the output bj

PR : a multivariate PolynomialRing defining the type of the expansion

ExpandLGF.far_field_expansion_coeffs — Function

far_field_expansion_coeffs(stencil::Stencil{3, T}, m, PR::MPolyRing{T}) where T

Calculates an $m$-term far-field expansion of the LGF using an algorithm from Martinsson and Rodin. The result is an array g of multivariate polynomials in $n$ so that

\[G(n) = \frac{1}{4\pi |\bm{n}|} + \frac{1}{\pi}\sum_{k=1}^m \frac{g_k(\bm{n})}{|\bm{n}|^{6k+1}} + \mathcal{O}\left( |\bm{n}|^{-(2m + 3)} \right)\]

Note that once $m$ exceeds 5 or 6, each additional term will roughly double the runtime of this function.

Arguments

stencil : the 3D finite difference stencil

m : number of correction terms returned

PR : a multivariate PolynomialRing that defines the type of the output

far_field_expansion_coeffs(stencil::Stencil{2, T}, m, PR::MPolyRing{T}) where T

Calculates an $m$-term far-field expansion of the 2D LGF using an algorithm from Martinsson and Rodin. The result is an array g of multivariate polynomials in $n$ so that

\[G(\bm{n}) = -\frac{1}{2\pi}\left(\log |\bm{n}| + \gamma + \frac{\log 8}{2}\right) + \frac{1}{\pi}\sum_{k=1}^m \frac{g_k(\bm{n})}{|\bm{n}|^{6j}} + \mathcal{O}\left( |\bm{n}|^{-(m+1)} \right),\]

Note that once $m$ exceeds 5 or 6, each additional term will roughly double the runtime of this function.

Arguments

stencil : the 2D finite difference stencil

m : number of correction terms returned

PR : a multivariate PolynomialRing that defines the type of the output

ExpandLGF.martinsson_rodin_a2q — Function

a2q = martinsson_rodin_a2q(stencil::Stencil, q_max::Int, PR::MPolyRing)

Return the polynomials $a_{2q}(\bm{k})$ with q <= q_max for a given stencil.

See Martinsson, Rodin. "Asymptotic Expansions of Lattice Green's Functions". JCP 2022

Arguments

cffs : the finite difference coefficients

q_max : the first q_max polynomials are returned

PR : a vector of 3 generators from a multivariate PolynomialRing

Callable expansions

ExpandLGF.inner_integral_expansion — Function

inner_integral_expansion(stencil::SplitStencil, terms)

Return a function which evaluates $I_{\sigma,n}(t)$ for large $t$ via a series expansion.

Arguments

stencil : coefficients of the FD scheme

terms : the number of terms in the expansion

Return

out_fun : a scalar function with signature val = out_fun(n, t)

inner_integral_expansion(b)

Return a function which evaluates $I_{\sigma,n}(t)$ for large $t$ via a series expansion.

Arguments

b : expansion coefficients, returned by inner_integral_expansion_coeffs

Return

out_fun : a scalar function with signature val = out_fun(n, t)

ExpandLGF.outer_integral_expansion — Function

outer_integral_expansion(stencil::SplitStencil{N, T}, terms) where {N, T}

Return a Julia function which evaluates the outer integral over $[t,\,\infty]$ for large $t$ via a series expansion.

Arguments

stencil : coefficients of the FD scheme

terms : the number of terms in the expansion

Return

out_fun : a scalar function with signature val = out_fun(n, t) for an N-Vector n and scalar t.

ExpandLGF.outer_integral_2d_expansion — Function

outer_integral_2d_expansion(g)

Return a Julia function which evaluates the 2D outer integral over $[T,\,\infty]$ for large $T$ via a series expansion.

Arguments

g : expansion coefficients, the result of outer_integral_2d_expansion_coeffs()

Return

out_fun : a scalar function with signature val = out_fun([n1, n2], T)

ExpandLGF.outer_integral_3d_expansion — Function

outer_integral_3d_expansion(g)

Return a Julia function which evaluates the 3D outer integral over $[T,\,\infty]$ for large $T$ via a series expansion.

Arguments

g : expansion coefficients, the result of outer_integral_3d_expansion_coeffs()

Return

out_fun : a scalar function with signature val = out_fun([n1, n2, n3], T)

ExpandLGF.far_field_expansion — Function

far_field_expansion(stencil::Stencil{N, T}, nterms, target = Symbolics.JuliaTarget()) where {N, T}

Return a function which evaluates a far-field expansion of the LGF.

When target = Symbolics.JuliaTarget() (default), outputs a function G = out_fun(n) for scalar G, N-vector n. When target = Symbolics.CTarget(), outputs a string containing a C function that evaluates the expansion.

Arguments

stencil : coefficients of the finite difference stencil

nterms : number of terms in the expansion

target : choose either Symbolics.JuliaTarget()` or Symbolics.CTarget()

ExpandLGF.far_field_2d_expansion — Function

far_field_2d_expansion(g, target::Symbolics.JuliaTarget)

Return a Julia function which evaluates a far-field expansion of the 2D LGF with coefficients in g. The output has signature G = out_fun(n) for scalar G, 2-vector n

Arguments

g : expansion coefficients, returned by far_field_2d_expansion_coeffs

far_field_2d_expansion(g, target::Symbolics.CTarget)

Return a string containing C code that evaluates a far-field expansion of the 2D LGF.

Arguments

g : expansion coefficients, returned by far_field_2d_expansion_coeffs

ExpandLGF.far_field_3d_expansion — Function

far_field_3d_expansion(g, target::JuliaTarget)

Return a Julia function which evaluates a far-field expansion of the 3D LGF. The output has signature G = out_fun(n) for scalar G, 3-vector n

Arguments

g : expansion coefficients, returned by far_field_3d_expansion_coeffs

far_field_3d_expansion(g, target::CTarget)

Return a string containing C code that evaluates a far-field expansion of the 3D LGF.

Arguments

g : far-field expansion coefficients, returned by far_field_3d_expansion_coeffs

Error estimates

ExpandLGF.inner_integral_errors — Function

tmax = inner_integral_errors(b, nmax; rtol=1e-16, atol=1e-16)

Calculate the region of validity for an inner integral expansion.

Specifically, when $n \le n_{\mathrm{max}}$ and $t \ge t_{\mathrm{max}}$, the last term in the expansion is smaller than both atol and rtol * (first term).

Arguments

b : vector of expansion coefficients (polynomials in n)

nmax : the maximal value of n that will be used in the expanions

rtol : relative tolerance on the last term in the series (pass 0 to ignore)

atol : absolute tolerance on the last term in the series (pass 0 to ignore)

ExpandLGF.outer_integral_errors — Function

Tmax = outer_integral_errors(g, nmax; rtol=1e-16, atol=1e-16)

Calculate the region of validity for an outer integral expansion. The dimension (2D or 3D) is inferred from the type of g.

Specifically, when $\|n\|_\infty ≤ n_{\mathrm{max}}$ and $T \ge T_{\mathrm{max}}$, the last term in the expansion is smaller than both atol and rtol * (first term).

Arguments

g : vector of expansion coefficients (polynomials in n1, n2, n3)

nmax : the maximal value of the 3-vector n that will be used in the expanions

rtol : relative tolerance on the last term in the series (pass 0 to ignore)

atol : absolute tolerance on the last term in the series (pass 0 to ignore)

ExpandLGF.far_field_errors — Function

nmin = far_field_errors(g; rtol=1e-16, atol=1e-16, search=100)

Calculate the region of validity for a far-field expansion. The dimension (2D or 3D) is inferred from the type of g.

Specifically, when $\|n\|_2 \ge n_{\mathrm{min}}$, the last term in the expansion is smaller than both atol and rtol * (first term).

Arguments

g : vector of far-field expansion coefficients (polynomials in n1, n2, n3)

rtol : relative tolerance on the last term in the series (pass 0 to ignore)

atol : absolute tolerance on the last term in the series (pass 0 to ignore)

search : finding nmin requires a bisection search over n. The search arg provides an upper limit.